requirements.txt
# Local scripts — ingestion, loading, transformation, quality
# Airflow and Kafka run in Docker and do NOT go here

yfinance>=0.2.40
pandas==2.0.0
numpy==1.24.0
kafka-python==2.0.2
dbt-snowflake==1.7.0
great-expectations==0.18.0
snowflake-connector-python==3.1.0
snowflake-sqlalchemy==1.5.1
python-dotenv==1.0.0
pydantic>=1.10.0,<2.0.0
requests==2.31.0
